Thesis proposal
I am currently writing my master thesis at Chalmers in the area of Interactive Design. The working title is Testing cultural sensitivity of mobile application for students' self help.

The main problem to be addressed in this project is the lack of application services for the student population. In addition to studying, students often pursue a hobby or enroll in additional courses. The question is why they could not share their knowledge or skills with other students?